Replace the RFC822 datetime parser; update tests
Much of the _parse_date_rfc822() code was copied from the rfc822
module back in 2005 and needed to be ripped out. Replacing the code
fixes what appears to be a copyright infringement issue and also will
let me get rid of the dependence on the rfc822 module.

What's lost with this change is out-of-range rollovers (e.g. Jan 32).
I don't consider this to be a significant issue, however.

git-svn-id: http://feedparser.googlecode.com/svn/trunk@683 73d2b349-402e-0410-baf4-070fd12ab5b7