Provide a fallback from `updated` to `published`
Fixes issue 328.

This is a necessary, though temporary, hack that will allow applications to
continue relying on feedparser's old broken behavior of saving the RSS element
pubDate in `updated`. This was fixed in issue 310, but old code may have
depended on this broken behavior, so it's necessary to allow developers time to
adapt their code and release new versions, and for users to upgrade their
software. This hack will be removed in a future version of feedparser.

Thanks to Thomas Perl for catching this and proposing a fix!

git-svn-id: http://feedparser.googlecode.com/svn/trunk@689 73d2b349-402e-0410-baf4-070fd12ab5b7